requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
dataclasses==0.6
dnspython==2.0.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
future==0.18.2
html5lib==1.1
idna==2.10
install==1.3.4
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.2
joblib==0.17.0
lxml==4.6.2
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.4
packaging==20.7
pandas==1.1.4
Pillow==8.0.1
protobuf==3.14.0
psycopg2==2.8.6
pymongo==3.11.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.24.0
requests-file==1.5.1
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
sentencepiece==0.1.91
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tokenizers==0.9.3
torch==1.7.0
tqdm==4.53.0
transformers==3.5.1
typing-extensions==3.7.4.3
urllib3==1.25.11
vaderSentiment==3.3.2
webencodings==0.5.1
Werkzeug==1.0.1
